refactor: rename Units folder to units
